Update requirements.txt
Add pyLDAvis to requirements and fix scikit-learn version to 1.1.3.

Fixing scikit-learn to this version is a temporary fix to allow get_feature_names and get_feature_names_out to both work for CountVectorizer in scikit-learn.

get_feature_names was the command used in the version of scikit-learn used during development of this project, however get_feature_names_out is required by pyLDAvis.